后台用户列表

+++ src/middleware/authorize/service.rs
@@ -84,7 +84,7 @@ where
         // query database
         let query_db_fut =
             Box::pin(UserSqlOperate::find_user_with_version_verify(
-                user.id as i64,
+                user.id,
                 user.num_pwd_change,
                 |user| user,
                 AuthorizeError::TOkenInvalid,
